+++ CHANGELOG.md
@@ -5,18 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [2.6.13] - 2021-07-29
 ### Fixed
 - Fixed miscelanious issue related to analytics logging [#638](https://github.com/rokwire/illinois-app/issues/638).
 - Crash when editing group event [#643](https://github.com/rokwire/illinois-app/issues/643).
 ### Changed
 - Show public/private event in Event detail panel [#622](https://github.com/rokwire/illinois-app/issues/622).
 - Pop to Group Detail Panel when user replies on a post [#623](https://github.com/rokwire/illinois-app/issues/623).
+- Add the posts bar with the + sign for members and admins even if there are no posts [#649](https://github.com/rokwire/illinois-app/issues/649).
 ### Added
 - Add bullets in group "About" section [#645](https://github.com/rokwire/illinois-app/issues/645).
 ### Fixed
 - Updated group privacy descriptions [#647](https://github.com/rokwire/illinois-app/issues/647).
 
-## [2.6.12] - 2021-07-27
+## [2.6.12] - 2021-07-28
 ### Fixed
 - Require user to input event end date for all day events [#627](https://github.com/rokwire/illinois-app/issues/627).
 - Event display date time [#626](https://github.com/rokwire/illinois-app/issues/626).